backend/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
databases==0.8.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==20.1.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
jose==1.0.0
Mako==1.3.0
MarkupSafe==2.1.3
mirakuru==2.5.2
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
port-for==0.7.2
psutil==5.9.6
psycopg==3.1.13
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
zipp==3.17.0
